TXRH-5661 undo the config feature change and add in much simpler overrides arguments to log func
The previous feature with configurations was not usable because the integrating apps interact with this library through static methods e.g. static call `mParticleCapacitorWeb.mParticleInit()`, so therefore calling the setter `setMParticleCapacitorConfiguration` wouldn't work because it would try to save the new config to `this.mparticleCapacitorConfiguration` which wouldn't be retrieved when integrating app calls the static method `mParticleInit()`.

The new approach is simpler and just allows the integrating app to supply in overriding values to modify the behavior of the logging functions.  Configurations is reverted back to a single config object built by `mParticleConfig()` and in that we attempt to determine defaults if any.

Sandbox application and unit tests have been changed to account for this rewriting of the feature.