TXRH-5661 fix and correct the return types
mparticle web-sdk does not expose any kind of callback or promise for us to use.  The log functions in web-sdk are fire and forget and so we should be making it look like there's a promise we hook onto and listen, because that's not the reality.

But in the login and logout functions, there is a callback exposed, so we return a promise to utilize that callback.

This is a larger change and may be a breaking change.